test/backend/requirements.txt
django
psycopg2
psycopg2-binary
requests
django-environ
djangorestframework
python-decouple
djangorestframework-simplejwt
pyjwt
pillow
django-cors-headers
pyotp
qrcode
channels_redis==4.2.1
channels==4.2.0
daphne
watchfiles